chore(middleware): move ui-options to external script (#5684)
* chore(middleware): move ui-options to external script

* remove JSDOM, fix basename

* update smoke test

* add name for action

* set no cache header

* argh

* Change default language from 'es-US' to 'en-US'

* Update render.spec.ts

---------
