+++ .changeset/soft-cats-sip.md
@@ -1,0 +1,9 @@
+---
+'@verdaccio/middleware': patch
+'@verdaccio/core': patch
+'@verdaccio/server': patch
+'@verdaccio/api': patch
+'@verdaccio/web': patch
+---
+
+chore(middleware): move ui-options to external script
