+++ packages/middleware/test/render.spec.ts
@@ -1,5 +1,4 @@
 import express from 'express';
-import { JSDOM } from 'jsdom';
 import path from 'node:path';
 import supertest from 'supertest';
 import { describe, expect, test } from 'vitest';
@@ -31,11 +30,13 @@ describe('test web server', () => {
     describe('output', () => {
       const render = async (config = 'default-test.yaml') => {
         const response = await supertest(initializeServer(config))
-          .get('/')
-          .set('Accept', HEADERS.TEXT_HTML)
-          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.TEXT_HTML_UTF8)
+          .get('/-/static/ui-options.js')
+          .set('Accept', HEADERS.JAVASCRIPT_CHARSET)
+          .expect(HEADERS.CACHE_CONTROL, HEADERS.NO_CACHE)
+          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.JAVASCRIPT_CHARSET)
           .expect(HTTP_STATUS.OK);
-        return new JSDOM(response.text, { runScripts: 'dangerously' });
+        const options = JSON.parse(response.text.slice(response.text.indexOf('=') + 1, -1));
+        return { window: { __VERDACCIO_BASENAME_UI_OPTIONS: options } };
       };
 
       test('should match render set ui properties', async () => {
@@ -63,7 +64,7 @@ describe('test web server', () => {
             pkgManagers: ['pnpm', 'yarn'],
             title: 'verdaccio web',
             scope: '@scope',
-            language: 'es-US',
+            language: 'en-US',
           })
         );
       });
@@ -103,13 +104,14 @@ describe('test web server', () => {
 
       test('should render successfully when request has no host header', async () => {
         const response = await supertest(initializeServer('default-test.yaml'))
-          .get('/')
-          .set('Accept', HEADERS.TEXT_HTML)
+          .get('/-/static/ui-options.js')
+          .set('Accept', HEADERS.JAVASCRIPT_CHARSET)
           .unset('Host')
-          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.TEXT_HTML_UTF8)
+          .expect(HEADERS.CACHE_CONTROL, HEADERS.NO_CACHE)
+          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.JAVASCRIPT_CHARSET)
           .expect(HTTP_STATUS.OK);
-        const dom = new JSDOM(response.text, { runScripts: 'dangerously' });
-        expect(dom.window.__VERDACCIO_BASENAME_UI_OPTIONS.basename).toEqual('/');
+        const options = JSON.parse(response.text.slice(response.text.indexOf('=') + 1, -1));
+        expect(options.basename).toEqual('/');
       });
 
       test.todo('should default title');
